修改：     core/ensemble_sample.hpp 	修改：     core/ptsample.hpp 	修改：     test_suit/test_coal.cpp

+++ test_suit/test_coal.cpp
@@ -95,7 +95,7 @@ int main()
   std::vector<double> beta_list;
   for(int i=0;i<ntemp;++i)
     {
-      beta_list.push_back(i/(ntemp-1));
+      beta_list.push_back(i/double(ntemp-1));
       std::vector<std::vector<double> > ensemble;
       for(int j=0;j<nwalker;++j)
 	{
